cleaned up minor annoyance
